pyproject.toml
[project]
name = "pdf-crawler"
version = "0.1.0"
description = "Universal Enterprise PDF Crawler"
requires-python = ">=3.10"
readme = "README.md"
authors = [
  { name = "AutoGenerated" }
]
dependencies = [
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "tqdm>=4.66",
  "playwright>=1.43",
  "pypdf>=4.0",
  "python-dotenv>=1.0"
]

[project.scripts]
pdf-crawler = "pdf_crawler.cli:main"

[tool.playwright]
install_dependencies = false
